Create a force-install script ...
This will fix the issue where we don't have the gems of the `ci-scripts`
group after running bundle-install for package construction.

This is a shame. We should really avoid depending on such things.

Should probably remove the `.rb` files from this.

Gotta think about the viability of using only shell + jq to do all of
the build tasks